requirements.txt
pyrogram 
qrcode
Pillow
